修改主键类型

+++ src/OnlineTestForCLanguage.Application/Exams/IExamsAppService.cs
@@ -7,7 +7,7 @@ using OnlineTestForCLanguage.Sessions.Dto;
 
 namespace OnlineTestForCLanguage.Sessions
 {
-    public interface IExamsAppService : IAsyncCrudAppService<ExamDto, int, PagedExamResultRequestDto, CreateExamDto, ExamDto>
+    public interface IExamsAppService : IAsyncCrudAppService<ExamDto, long, PagedExamResultRequestDto, CreateExamDto, ExamDto>
     {
     }
 }
